Change order of brewfile
Firstly these brewfiles are not idempotent, order is important, it's
like doing a brew install/tap/cask install/mas for each item in the
order it appears.

Secondly formula can't depend on casks.

Thirdly casks are relatively stand alone pkg files or similar.

That means we should be safest installing in the order

taps <- So we can access all the packages we need
cask <- So we have anything we depend on but can't represent in a formula
brew <- "Everything else"
mas  <- we install mas via brew